Attempting fix

+++ src/main/java/com/thebubblenetwork/api/game/spectator/PlayersList.java
@@ -91,7 +91,7 @@ public class PlayersList extends Menu implements Listener {
     protected ItemStack generate(Player p) {
         ItemStack is = generateSkull(p.getName());
         withName(is, ChatColor.AQUA + p.getName());
-        withLore(is, ChatColor.DARK_RED + "Health: " + ChatColor.GRAY + health.format(p.getHealth()) + "/" + String.valueOf(p.getMaxHealth()), ChatColor.YELLOW + "Hunger: " + ChatColor.GRAY + String.valueOf(p.getFoodLevel()) + "/20", "", ChatColor.DARK_PURPLE + "Left Click -> Spectate", ChatColor.DARK_AQUA + "Right Click -> View inventory");
+        withLore(is, ChatColor.DARK_RED + "Health: " + ChatColor.GRAY + health.format(p.getHealth()) + "/" + String.valueOf((int)p.getMaxHealth()), ChatColor.YELLOW + "Hunger: " + ChatColor.GRAY + String.valueOf(p.getFoodLevel()) + "/20", "", ChatColor.DARK_PURPLE + "Left Click -> Spectate", ChatColor.DARK_AQUA + "Right Click -> View inventory");
         return is;
     }
